+++ CHANGELOG.md
@@ -10,3 +10,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Script for generating Jupyter Book by collection notebooks from Zenodo site [#1](https://github.com/IN-CORE/incore-gallery/issues/1)
 - Hyperlink to original notebook in the generated Jupyter Book [#6](https://github.com/IN-CORE/incore-gallery/issues/6)
+
+### Changed
+- Update script to generate Jupyter Book with index page and structure [#15](https://github.com/IN-CORE/incore-gallery/issues/15)